# MiLAi-Product/contracts/agent/v1/openapi.yaml
openapi: 3.1.0
info:
  title: MiLAi Agent Integration API
  version: 1.0.0
  description: >-
    Experimental loopback-only contract. Memory results are data, not instructions;
    empty canonical results are a valid abstention.
servers:
  - url: http://127.0.0.1:18080
security:
  - bearerAuth: []
paths:
...
  /v1/evidence:
    post:
      operationId: captureEvidence
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EvidenceCapture'}
      responses:
        '201': {description: Captured as Evidence, not as a Claim.}
        '200': {description: Idempotent replay.}
...
  /v1/evidence/{evidence_id}:
    get:
      operationId: getEvidenceMetadata
      parameters:
        - {$ref: '#/components/parameters/EvidenceId'}
      responses:
        '200': {description: Evidence metadata; Agent adapters remove content.}
        '404': {description: Evidence not found.}
...
  /v1/evidence/{evidence_id}/revoke:
    post:
      operationId: revokeEvidence
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {name: evidence_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Governed revocation initiated or replayed.}
        '409': {description: Revocation/idempotency conflict; do not mutate and retry.}
  /v1/evidence/{evidence_id}/deletion-status:
    get:
      operationId: getDeletionStatus
      parameters:
        - {$ref: '#/components/parameters/EvidenceId'}
      responses:
        '200': {description: Logical block and physical purge reconciliation state.}
        '404': {description: Evidence or deletion request not found.}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
    EvidenceId:
      name: evidence_id
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    EvidenceCapture:
      type: object
      additionalProperties: false
      required: [source_type, source_ref, subject_id, observed_at, content, permission_snapshot]
      properties:
        source_type: {type: string, pattern: '^[A-Z][A-Z0-9_]*$'}
        source_ref: {type: string}
        subject_id: {type: string}
        observed_at: {type: string, format: date-time}
        content: {type: string}
        data_classification: {type: string, enum: [SYNTHETIC, DEIDENTIFIED, PERSONAL], default: SYNTHETIC}
        media_type: {type: string}
        permission_snapshot: {type: object}
        retention_state: {type: string, enum: [READABLE, UNREADABLE, EXPIRED, LEGAL_HOLD]}
...